+++ cli/pull.ts
@@ -1,16 +1,6 @@
 // deno-lint-ignore-file no-explicit-any
-import { resolveWorkspace } from "./context.ts";
 import { GlobalOptions } from "./types.ts";
-import {
-  colors,
-  Command,
-  Confirm,
-  copy,
-  ensureDir,
-  path,
-  readerFromStreamReader,
-  Untar,
-} from "./deps.ts";
+import { colors, Command, readerFromStreamReader, Untar } from "./deps.ts";
 import { Workspace } from "./workspace.ts";
 
 export async function downloadTar(
@@ -50,46 +40,15 @@ export async function downloadTar(
   return untar;
 }
 
-async function pull(opts: GlobalOptions & { override: boolean }, dir: string) {
-  const workspace = await resolveWorkspace(opts);
-
-  const untar = await downloadTar(workspace);
-  if (!untar) return;
-
-  for await (const entry of untar) {
-    console.log(entry.fileName);
-    const filePath = path.resolve(dir, entry.fileName);
-    if (entry.type === "directory") {
-      await ensureDir(filePath);
-      continue;
-    }
-    await ensureDir(path.dirname(filePath));
-    if (!opts.override) {
-      let exists = false;
-      try {
-        const _stat = await Deno.stat(filePath);
-        exists = true;
-      } catch {
-        exists = false;
-      }
-      if (exists) {
-        if (
-          !(await Confirm.prompt(
-            "Conflict at " +
-              filePath +
-              " do you want to override the local version?",
-          ))
-        ) {
-          continue;
-        }
-      }
-    }
-    const file = await Deno.open(filePath, { write: true, create: true });
-    const len = await copy(entry, file);
-    await file.truncate(len);
-    file.close();
-  }
-  console.log(colors.green("Done. Wrote all files to disk."));
+async function stub(
+  _opts: GlobalOptions & { override: boolean },
+  _dir: string,
+) {
+  console.log(
+    colors.red.underline(
+      'Pull is deprecated. Use "sync pull --raw" instead. See <TODO_LINK_HERE> for more information.',
+    ),
+  );
 }
 
 const command = new Command()
@@ -97,6 +56,6 @@ const command = new Command()
     "Pull all definitions in the current workspace from the API and write them to disk.",
   )
   .arguments("<dir:string>")
-  .action(pull as any);
+  .action(stub as any);
 
 export default command;
