+++ cli/sync.ts
@@ -5,6 +5,8 @@ import {
   cbor,
   colors,
   Command,
+  Confirm,
+  copy,
   ensureDir,
   iterateReader,
   microdiff,
@@ -20,6 +22,8 @@ import {
 } from "./types.ts";
 import { downloadTar } from "./pull.ts";
 import { FolderFile } from "./folder.ts";
+import { ResourceTypeFile } from "./resource-type.ts";
+import { findContentFile, pushScript, ScriptFile } from "./script.ts";
 
 type TrackedId = string;
 const TrackedId = String;
@@ -225,7 +229,18 @@ async function updateStateFromRemote(
   await state.save();
 }
 
-async function pull(opts: GlobalOptions) {
+async function pull(
+  opts: GlobalOptions & { raw: boolean; rawOverride: boolean },
+) {
+  if (opts.raw) {
+    const opts2 = opts as any;
+    opts2.override = opts2.rawOverride;
+    opts2.raw = undefined;
+    opts2.rawOverride = undefined;
+    await pullRaw(opts2, Deno.cwd());
+    return;
+  }
+
   const state = await getState(opts);
 
   const workspace = await resolveWorkspace(opts);
@@ -309,7 +324,14 @@ async function pull(opts: GlobalOptions) {
   }
 }
 
-async function push(opts: GlobalOptions) {
+async function push(opts: GlobalOptions & { raw: boolean }) {
+  if (opts.raw) {
+    const opts2 = opts as any;
+    opts2.raw = undefined;
+    await pushRaw(opts2, undefined);
+    return;
+  }
+
   const state = await getState(opts);
 
   const workspace = await resolveWorkspace(opts);
@@ -391,9 +413,6 @@ async function add(opts: GlobalOptions, path: string) {
   await state.save();
 }
 
-async function sync(opts: GlobalOptions) {
-}
-
 async function init(opts: GlobalOptions) {
   const root = Deno.cwd();
   try {
@@ -421,9 +440,307 @@ async function init(opts: GlobalOptions) {
   await newState.save();
 }
 
+async function pullRaw(
+  opts: GlobalOptions & { override: boolean },
+  dir: string,
+) {
+  const workspace = await resolveWorkspace(opts);
+
+  const untar = await downloadTar(workspace);
+  if (!untar) return;
+
+  for await (const entry of untar) {
+    console.log(entry.fileName);
+    const filePath = path.resolve(dir, entry.fileName);
+    if (entry.type === "directory") {
+      await ensureDir(filePath);
+      continue;
+    }
+    await ensureDir(path.dirname(filePath));
+    if (!opts.override) {
+      let exists = false;
+      try {
+        const _stat = await Deno.stat(filePath);
+        exists = true;
+      } catch {
+        exists = false;
+      }
+      if (exists) {
+        if (
+          !(await Confirm.prompt(
+            "Conflict at " +
+              filePath +
+              " do you want to override the local version?",
+          ))
+        ) {
+          continue;
+        }
+      }
+    }
+    const file = await Deno.open(filePath, { write: true, create: true });
+    const len = await copy(entry, file);
+    await file.truncate(len);
+    file.close();
+  }
+  console.log(colors.green("Done. Wrote all files to disk."));
+}
+
+type PushRawCandidate = {
+  path: string;
+  namespaceKind: "user" | "group" | "folder";
+  namespaceName: string;
+};
+
+type PushRawResourceTypeCandidate = {
+  path: string;
+};
+
+type PushRawFolderCandidate = {
+  path: string;
+  namespaceName: string;
+};
+
+async function pushRawFindCandidateFiles(
+  dir: string,
+): Promise<
+  {
+    normal: PushRawCandidate[];
+    resourceTypes: PushRawResourceTypeCandidate[];
+    folders: PushRawFolderCandidate[];
+  }
+> {
+  dir = path.resolve(dir);
+  if (path.dirname(dir).startsWith(".")) {
+    return { normal: [], resourceTypes: [], folders: [] };
+  }
+  const normalCandidates: PushRawCandidate[] = [];
+  const resourceTypeCandidates: PushRawResourceTypeCandidate[] = [];
+  const folderCandidates: PushRawFolderCandidate[] = [];
+  for await (const e of Deno.readDir(dir)) {
+    if (e.isDirectory) {
+      if (e.name == "u" || e.name == "g" || e.name == "f") { // TODO: Check version for f
+        const newDir = dir + (dir.endsWith("/") ? "" : "/") + e.name;
+        for await (const e2 of Deno.readDir(newDir)) {
+          if (e2.isDirectory) {
+            if (e2.name.startsWith(".")) continue;
+            const namespaceName = e2.name;
+            const stack: string[] = [];
+            {
+              const path = newDir + "/" + namespaceName + "/";
+              stack.push(path);
+              try {
+                await Deno.stat(path + "folder.meta.json");
+                folderCandidates.push({
+                  namespaceName,
+                  path: path + "folder.meta.json",
+                });
+              } catch {}
+            }
+
+            while (stack.length > 0) {
+              const dir2 = stack.pop()!;
+              for await (const e3 of Deno.readDir(dir2)) {
+                if (e3.isFile) {
+                  if (e3.name === "folder.meta.json") continue;
+                  normalCandidates.push({
+                    path: dir2 + e3.name,
+                    namespaceKind: e.name == "g"
+                      ? "group"
+                      : e.name == "u"
+                      ? "user"
+                      : "folder",
+                    namespaceName: namespaceName,
+                  });
+                } else {
+                  stack.push(dir2 + e3.name + "/");
+                }
+              }
+            }
+          }
+        }
+      } else {
+        console.log(
+          colors.yellow(
+            "Including organizational folder " + e.name + " in push!",
+          ),
+        );
+        const { normal, resourceTypes, folders } =
+          await pushRawFindCandidateFiles(
+            path.join(dir, e.name),
+          );
+        normalCandidates.push(...normal);
+        resourceTypeCandidates.push(...resourceTypes);
+        folderCandidates.push(...folders);
+      }
+    } else {
+      // handle root files
+      if (e.name.endsWith(".resource-type.json")) {
+        resourceTypeCandidates.push({
+          path: dir + (dir.endsWith("/") ? "" : "/") + e.name,
+        });
+      }
+    }
+  }
+  return {
+    normal: normalCandidates,
+    folders: folderCandidates,
+    resourceTypes: resourceTypeCandidates,
+  };
+}
+
+async function pushRaw(opts: GlobalOptions, dir?: string) {
+  dir = dir ?? Deno.cwd();
+  const workspace = await resolveWorkspace(opts);
+  await requireLogin(opts);
+
+  console.log(colors.blue("Searching Directory..."));
+  const { normal, resourceTypes, folders } = await pushRawFindCandidateFiles(
+    dir,
+  );
+  console.log(
+    colors.blue(
+      "Found " + (normal.length + resourceTypes.length + folders.length) +
+        " candidates",
+    ),
+  );
+  for (const resourceType of resourceTypes) {
+    const fileName = resourceType.path.substring(
+      resourceType.path.lastIndexOf("/") + 1,
+    );
+    const fileNameParts = fileName.split(".");
+    // invalid file names, like my.cool.script.script.json. Not valid.
+    if (fileNameParts.length != 3) {
+      console.log(
+        colors.yellow("invalid file name found at " + resourceType.path),
+      );
+      continue;
+    }
+
+    // filter out non-json files. Note that we filter out script contents above, so this is really an error.
+    if (fileNameParts.at(-1) != "json") {
+      console.log(colors.yellow("non-JSON file found at " + resourceType.path));
+      continue;
+    }
+
+    console.log("pushing resource type " + fileNameParts.at(-3)!);
+    await decoverto.type(ResourceTypeFile).rawToInstance(
+      await Deno.readTextFile(resourceType.path),
+    ).push(workspace.workspaceId, fileNameParts.at(-3)!);
+  }
+  for (const folder of folders) {
+    await decoverto.type(FolderFile).plainToInstance(
+      JSON.parse(await Deno.readTextFile(folder.path)),
+    ).push(
+      workspace.workspaceId,
+      "f/" + folder.namespaceName,
+    );
+  }
+  for (const candidate of normal) {
+    // full file name. No leading /. includes .type.json
+    const fileName = candidate.path.substring(
+      candidate.path.lastIndexOf("/") + 1,
+    );
+    // figure out just the path after ...../u|g/username|group/ (in extra dir)
+    const dirParts = candidate.path.split("/").filter((x) => x.length > 0);
+    // TODO: check version for folder
+    const gIndex = dirParts.findIndex((x) => x == "u" || x == "g" || x == "f");
+    const extraDir = dirParts.slice(gIndex + 2, -1).join("/");
+
+    // file name parts has .json (hopefully) at -1, type at -2, and the actual name at -3. Dots in names are not allowed.
+    const fileNameParts = fileName.split(".");
+
+    // filter out script content files
+    if (
+      fileNameParts.at(-1) == "ts" ||
+      fileNameParts.at(-1) == "py" ||
+      fileNameParts.at(-1) == "go"
+    ) {
+      // probably part of a script. Silent ignore.
+      continue;
+    }
+
+    // invalid file names, like my.cool.script.script.json. Not valid.
+    if (fileNameParts.length != 3) {
+      console.log(
+        colors.yellow("invalid file name found at " + candidate.path),
+      );
+      continue;
+    }
+
+    // filter out non-json files. Note that we filter out script contents above, so this is really an error.
+    if (fileNameParts.at(-1) != "json") {
+      console.log(colors.yellow("non-JSON file found at " + candidate.path));
+      continue;
+    }
+
+    // get the type & filter it for valid ones.
+    const type = fileNameParts.at(-2);
+    if (type == "resource-type") {
+      console.log(
+        colors.yellow(
+          "Found resource type file at " +
+            candidate.path +
+            " this appears to be inside a path folder. Resource types are not addressed by path. Place them at the root or inside only an organizational folder. Ignoring this file!",
+        ),
+      );
+      continue;
+    }
+
+    if (
+      type != "flow" &&
+      type != "resource" &&
+      type != "script" &&
+      type != "variable"
+    ) {
+      console.log(
+        colors.yellow(
+          "file with invalid type " + type + " found at " + candidate.path,
+        ),
+      );
+      continue;
+    }
+
+    // create the remotePath for the API
+    const remotePath = (candidate.namespaceKind === "group"
+      ? "g/"
+      : (candidate.namespaceKind === "user" ? "u/" : "f/")) +
+      candidate.namespaceName +
+      "/" +
+      (extraDir.length > 0 ? extraDir + "/" : "") +
+      fileNameParts.at(-3);
+
+    console.log("pushing " + type + " to " + remotePath);
+
+    const typed = inferTypeFromPath(
+      candidate.path,
+      JSON.parse(await Deno.readTextFile(candidate.path)),
+    );
+    if (typed instanceof ResourceTypeFile || typed instanceof FolderFile) {
+      throw new Error(
+        "Resource Types and Folders should  be filtered out at this point!",
+      );
+    } else if (typed instanceof ScriptFile) {
+      let contentPath: string;
+      try {
+        contentPath = await findContentFile(candidate.path);
+      } catch (e) {
+        console.log(colors.red(e.toString()));
+        continue;
+      }
+      await pushScript(
+        candidate.path,
+        contentPath,
+        workspace.workspaceId,
+        remotePath,
+      );
+    } else {
+      typed.push(workspace.workspaceId, remotePath);
+    }
+  }
+  console.log(colors.underline.bold.green("Successfully Pushed all files."));
+}
+
 const command = new Command()
-  .description("Sync the current folder")
-  .action(sync as any)
   .command("init")
   .description(
     "Initialize this folder as sync root for the currently selected workspace & remote.",
@@ -435,9 +752,14 @@ const command = new Command()
   .action(add as any)
   .command("pull")
   .description("Pull any remote changes and apply them locally")
+  .option("--raw", "Pull without using state.")
+  .option("--raw-override", "Always override local files with remote.", {
+    depends: ["raw"],
+  })
   .action(pull as any)
   .command("push")
   .description("Push any local changes and apply them remotely")
+  .option("--raw", "Push without using state.")
   .action(push as any);
 
 export default command;
